declare nuclide classification
