+++ enum/science/nuclide.ts
@@ -1,0 +1,22 @@
+export enum NuclideState {
+  GROUND = 'ground',
+  M1 = 'm1',
+  M2 = 'm2',
+  M3 = 'm3',
+  M4 = 'm4'
+};
+
+export enum NuclideStability {
+  STABLE = 'stable',
+  UNSTABLE = 'unstable'
+};
+
+export enum NuclideOrigin {
+  PRIMORDIAL = 'primordial',
+  NATURAL = 'natural',
+  SYNTHETIC = 'synthetic'
+};
+
+export enum NuclideProperty {
+  MONONUCLIDE = 'mononuclide'
+};
